move getWrappedTvslst() to shared/utils/src/filter.js
